people-search/Search{Form,Input}: restructure into own dirs
