--- src/components/people-search/Search/index.js
@@ -1,2 +1,0 @@
-export { default as SearchForm } from './Form';
-export { default as SearchInput } from './Input';
