+++ src/components/people-search/SearchForm/index.js
@@ -1,0 +1,2 @@
+import SearchForm from './SearchForm';
+export default SearchForm;
