+++ src/components/people-search/SearchInput/SearchInput.jsx
@@ -10,6 +10,6 @@ export default function SearchInput(props) {
 
     return (
         <View></View>
-    )
+    );
 
 }
