+++ src/components/people-search/SearchInput/SearchInput.styles.js
