+++ src/components/people-search/SearchInput/index.js
@@ -1,0 +1,2 @@
+import SearchInput from './SearchInput';
+export default SearchInput;
